requirements.txt
altgraph==0.17
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
comtypes==1.1.7
constantly==15.1.0
countries==0.2.0
cryptography==3.2.1
cssselect==1.1.0
dataclasses-json==0.5.2
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
django-channels==0.7.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-phone-field==1.8.1
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
future==0.18.2
futures==3.1.1
google-search==1.0.2
goslate==1.5.1
gsearch==1.6.0
gTTS==2.2.1
gunicorn==20.1.0
html5lib==1.1
hyperlink==20.0.1
idna==2.10
imutils==0.5.3
incremental==17.5.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==5.6.4
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
keyboard==0.13.5
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-enum==1.5.1
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
msgpack==1.0.0
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
openpyxl==3.0.5
packaging==20.8
pandocfilters==1.4.3
parse==1.19.0
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.1
playsound==1.2.2
prometheus-client==0.9.0
prompt-toolkit==3.0.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Tanika%20priya/downloads/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
PyDictionary==2.0.1
pyee==8.1.0
PyGetWindow==0.0.9
Pygments==2.7.3
PyHamcrest==2.0.2
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pylint==2.6.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.0
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
pyppeteer==0.2.5
pyquery==1.4.3
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
python-dateutil==2.8.1
python-youtube==0.6.4
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.1
pywhatkit==3.7
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.0
requests-html==0.10.0
requests-oauthlib==1.3.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
spotipy==2.16.1
sqlparse==0.4.1
stringcase==1.2.0
terminado==0.9.1
testpath==0.4.4
text-to-speech==1.3.3
toml==0.10.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
txaio==20.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
validate-email==1.3
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.12.1
zope.interface==5.2.0
